requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
qdrant-client>=1.7.0
sentence-transformers>=2.3.0
PyPDF2>=3.0.0
python-docx>=1.1.0
tiktoken>=0.5.0
rank-bm25>=0.2.2
numpy>=1.24.0
